API Key Manager Added
Substantial changes to centralize the management of API keys.  These are critical for security, and also we'll want to extend our use of keys to support more than one in the app.  This might be different keys for different uses (e.g. main page and test), but also other keys (e.g. Gemini or Anthropic).

This puts in place the foundation for much cleaner, safer, flexible handling.